flask_backend/requirements.txt
Flask
psycopg2-binary
gunicorn
python-dotenv
flask-cors
bcrypt
Flask-JWT-Extended